+++ pages/search.tsx
@@ -2,11 +2,7 @@ import { GetServerSideProps, NextPage } from "next";
 import Head from "../components/head";
 import Header from "../components/header";
 import RecipeList from "../components/recipe-list";
-import {
-  Recipe,
-  searchRecipes,
-  SearchRecipesResponse,
-} from "../lib/client/recipe";
+import { Recipe, searchRecipes, SearchRecipesResponse } from "../lib/recipe";
 
 type Props = {
   // このページで表示するレシピのリスト
